Restructed lib into siloed modules including utils and cache manager
