+++ geocode/postcodes2latlon.py
@@ -3,6 +3,7 @@
 Load a list of postcodes from a CSV file and geocode them to latitude/longitude.
 
 - Jamie Taylor <jamie.taylor@sheffield.ac.uk>
+- Ethan Jones <ejones18@sheffield.ac.uk>
 - First Authored: 2020-06-12
 """
 
@@ -12,13 +13,14 @@ import argparse
 import time as TIME
 import pandas as pd
 
-from geocode import Geocoder, query_yes_no
+from geocode import Geocoder
+from utilities import query_yes_no, GenericException
 
 def parse_options():
     """Parse command line options."""
     parser = argparse.ArgumentParser(description=("This is a command line interface (CLI) for "
                                                   "the postcodes2latlon.py module"),
-                                     epilog="Jamie Taylor, 2020-06-12")
+                                     epilog="Jamie Taylor & Ethan Jones, 2020-06-12")
     parser.add_argument("-f", "--input-file", dest="infile", action="store", type=str,
                         required=True, metavar="</path/to/file>",
                         help="Specify a CSV file containing a list of postcodes to be geocoded. "
@@ -28,7 +30,7 @@ def parse_options():
                         required=True, metavar="</path/to/file>", help="Specify an output file.")
     options = parser.parse_args()
     if not os.path.isfile(options.infile):
-        raise Exception(f"The input file '{options.infile}' does not exist.")
+        raise GenericException(f"The input file '{options.infile}' does not exist.")
     if os.path.isfile(options.outfile):
         check = query_yes_no(f"The outfile '{options.outfile}' already exists, are you sure you "
                              "wish to overwrite?", "no")
@@ -42,8 +44,8 @@ def main():
     options = parse_options()
     with open(options.infile, "r") as fid:
         df = pd.read_csv(fid).iloc[:100]
-    with Geocoder(progress_bar=True) as geo:
-        results = geo.geocode(postcodes=df["postcode"].to_numpy())
+    with Geocoder() as geo:
+        results = geo.geocode_postcode(df["postcode"].to_numpy())
         lats = [r[0] if r[0] is not None else pd.NA for r in results]
         lons = [r[1] if r[1] is not None else pd.NA for r in results]
         status = [geo.status_codes[r[2]] for r in results]
